Support tmux
Supporting tmux required updating to the latest cview (v1.5.7) which
drew in changes in tcell.

In this commit we also move to cview's new repo, and finally remove the
import alias which aliased cview to tview.

Most the changes in `internal/gui` are what was required to compile
cave-logger on this newer cview, as there have been several breaking API
changes. The most affecting being the removinal of chaining. While
Trevor had good reasons for this it was a functionality that cave-logger
used extensively.

Also remove unused channel from page structs.